requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9

# Machine Learning & Computer Vision
torch==2.2.0
torchvision==0.17.0
ultralytics==8.1.0  # YOLOv8
opencv-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0
timm==0.9.12  # For MiDaS depth estimation
transformers==4.38.1  # For Hugging Face models

# Privacy
tensorflow-privacy==0.8.10
pydp==0.1.16  # Python Differential Privacy (updated version)

# Message Queue
kafka-python==2.0.2

# Database
redis==5.0.1
psycopg2-binary==2.9.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.5

# Utilities
python-dotenv==1.0.1
loguru==0.7.2 